feat(domain): add User model
This commit introduces the core User entity within the domain layer. This model represents a user of the system and contains their fundamental information.

The User model includes the following attributes:
- idUser (UUID)
- nombre
- apellido
- fechaNacimiento
- direccion
- telefono
- correoElectronico
- salarioBase

This entity is central to the business logic and will be used by use cases to perform operations related to users.